More files added into the split dirs
