+++ pgmpy/tests/test_estimators/test_ScoreCache.py
@@ -1,6 +1,6 @@
 import unittest
 from mock import Mock, MagicMock, call
-from pgmpy.estimators.ScoreCache import LRUCache, ScoreCache
+from pgmpy.structure_estimators.ScoreCache import LRUCache, ScoreCache
 from pgmpy.estimators import BIC
 import pandas as pd
